refactor(review): hide offset maps behind target-maps leaf
