+++ packages/lexical-review/src/ReviewTargeting.ts
@@ -12,7 +12,6 @@ import {
 } from "lexical";
 import {
   $isReviewFragmentNode,
-  $canReviewElementNodesBeMerged,
   $isReviewDeletionNode,
   $isReviewFormattingNode,
   $isReviewInsertionNode,
@@ -23,6 +22,18 @@ import {
   type ReviewElementNode,
   type ReviewFragmentNode,
 } from "./ReviewNodes";
+import {
+  buildAcceptedMap,
+  buildProposalMap,
+  getAcceptedOffset,
+  getProposalOffset,
+  isSameProposalNode,
+  isolateSpanNodes,
+  overlappingSlices,
+  proposalMapOf,
+  type AcceptedMap,
+  type ProposalMap,
+} from "./ReviewTargetMaps";
 import {
   refusal,
   type Preparation,
@@ -66,40 +77,10 @@ type SelectionInspection = Readonly<{
   selection: RangeSelection;
 }>;
 
-type OffsetEntry = Readonly<{
-  end: number;
-  node: TextNode;
-  start: number;
-}>;
-
-export type ProposalMapEntry = OffsetEntry &
-  Readonly<{
-    wrapper: ReviewElementNode;
-  }>;
-
-type ProposalMap = Readonly<{
-  entries: readonly ProposalMapEntry[];
-  paragraph: ParagraphNode;
-  total: number;
-  wrappers: readonly ReviewElementNode[];
-  proposalId: string;
-}>;
-
 type ProposalSpan = Readonly<{
   end: number;
   map: ProposalMap;
   start: number;
-}>;
-
-export type AcceptedMapEntry = OffsetEntry &
-  Readonly<{
-    childIndex: number;
-  }>;
-
-type AcceptedMap = Readonly<{
-  entries: readonly AcceptedMapEntry[];
-  paragraph: ParagraphNode;
-  total: number;
 }>;
 
 type AcceptedSpan = Readonly<{
@@ -121,17 +102,6 @@ function getRootProposalContext(
   }
   const paragraph = node.getParent();
   return isRootParagraph(paragraph) ? { paragraph, wrapper: node } : null;
-}
-
-function isSameProposalNode(
-  node: LexicalNode | null | undefined,
-  reference: ReviewElementNode,
-): node is ReviewElementNode {
-  return (
-    isReviewElementNode(node) &&
-    (node.getKey() === reference.getKey() ||
-      $canReviewElementNodesBeMerged(reference, node))
-  );
 }
 
 /** Paragraph placement rule shared by live targeting and serialized validation; unify the two checks once serialized validation can query this seam. */
@@ -747,108 +717,6 @@ export function inspectProposalGroup(proposalId: string): Preparation<{
   return validateCollectedProposalGroup(collectProposalNodes(proposalId));
 }
 
-type OffsetUnit<Taken> = Readonly<{
-  taken: Taken;
-  start: number;
-  end: number;
-}>;
-
-/**
- * One offset walk behind both associations. take admits each child (null
- * skips it); the loop stamps text offsets once for every admitted unit.
- */
-function collectOffsetUnits<Taken extends { node: TextNode }>(
-  children: readonly LexicalNode[],
-  startIndex: number,
-  endIndex: number,
-  take: (
-    child: LexicalNode | undefined,
-    childIndex: number,
-  ) => readonly Taken[] | null | ReviewIntentRefusal,
-): Preparation<{ units: Array<OffsetUnit<Taken>>; total: number }> {
-  const units: Array<OffsetUnit<Taken>> = [];
-  let offset = 0;
-  for (let index = startIndex; index <= endIndex; index += 1) {
-    const taken = take(children[index], index);
-    if (taken === null) {
-      continue;
-    }
-    if ("status" in taken) {
-      return taken;
-    }
-    for (const unit of taken) {
-      const end = offset + unit.node.getTextContentSize();
-      units.push({ taken: unit, start: offset, end });
-      offset = end;
-    }
-  }
-  return { status: "ready", value: { units, total: offset } };
-}
-
-function buildProposalMap(
-  paragraph: ParagraphNode,
-  startWrapper: ReviewElementNode,
-  endWrapper: ReviewElementNode,
-): Preparation<ProposalMap> {
-  const group = inspectProposalGroup(startWrapper.getProposalId());
-  if (group.status !== "ready") return group;
-  const startIndex = getChildIndex(paragraph, startWrapper);
-  const endIndex = getChildIndex(paragraph, endWrapper);
-  if (startIndex === null || endIndex === null || startIndex > endIndex) {
-    return refusal(
-      "invalid-structural-target",
-      "The proposal selection wrappers are not ordered in one paragraph.",
-    );
-  }
-  const proposalId = startWrapper.getProposalId();
-  const collected = collectOffsetUnits(
-    paragraph.getChildren(),
-    startIndex,
-    endIndex,
-    (child) => {
-      if (!isSameProposalNode(child, startWrapper)) {
-        return refusal(
-          "unsafe-proposal-intersection",
-          "The selection intersects accepted content or another proposal identity.",
-        );
-      }
-      const textChildren = getTextChildren(child);
-      if (textChildren === null) {
-        return refusal(
-          "invalid-structural-target",
-          "A pending proposal contains unsupported live children.",
-        );
-      }
-      return textChildren.map((node) => ({ node, wrapper: child }));
-    },
-  );
-  if (collected.status !== "ready") {
-    return collected;
-  }
-  const { units, total } = collected.value;
-  if (units.length === 0) {
-    return refusal(
-      "invalid-structural-target",
-      "A pending proposal must contain live text before it can be edited.",
-    );
-  }
-  return {
-    status: "ready",
-    value: {
-      entries: units.map(({ taken, start, end }) => ({
-        end,
-        node: taken.node,
-        start,
-        wrapper: taken.wrapper,
-      })),
-      paragraph,
-      proposalId,
-      total,
-      wrappers: [...new Set(units.map((unit) => unit.taken.wrapper))],
-    },
-  };
-}
-
 function buildProposalMapAroundPoint(
   point: ProposalPoint,
 ): Preparation<ProposalMap> {
@@ -876,37 +744,12 @@ function buildProposalMapAroundPoint(
       "The proposal caret is not attached to a supported proposal run.",
     );
   }
-  return buildProposalMap(point.paragraph, startWrapper, endWrapper);
-}
-
-export function getProposalOffset(
-  point: ProposalPoint,
-  map: ProposalMap,
-): number | null {
-  let offset = 0;
-  for (const wrapper of map.wrappers) {
-    const children = getTextChildren(wrapper);
-    if (children === null) {
-      return null;
-    }
-    if (wrapper.getKey() === point.wrapper.getKey()) {
-      if (point.node === null) {
-        return offset + point.offset;
-      }
-      for (const child of children) {
-        if (child.getKey() === point.node.getKey()) {
-          return offset + point.offset;
-        }
-        offset += child.getTextContentSize();
-      }
-      return null;
-    }
-    offset += children.reduce(
-      (total, child) => total + child.getTextContentSize(),
-      0,
-    );
-  }
-  return null;
+  return buildProposalMap(
+    point.paragraph,
+    startWrapper,
+    endWrapper,
+    inspectProposalGroup,
+  );
 }
 
 function buildProposalSpan(
@@ -933,6 +776,7 @@ function buildProposalSpan(
     startPoint.paragraph,
     startPoint.wrapper,
     endPoint.wrapper,
+    inspectProposalGroup,
   );
   if (map.status !== "ready") {
     return map;
@@ -946,66 +790,6 @@ function buildProposalSpan(
     );
   }
   return { status: "ready", value: { end, map: map.value, start } };
-}
-
-export function buildAcceptedMap(paragraph: ParagraphNode): AcceptedMap {
-  const children = paragraph.getChildren();
-  const collected = collectOffsetUnits(
-    children,
-    0,
-    children.length - 1,
-    (child, childIndex) =>
-      !$isTextNode(child) || child.getTextContentSize() === 0
-        ? null
-        : [{ node: child, childIndex }],
-  );
-  if (collected.status !== "ready") {
-    throw new Error("Accepted admission cannot refuse.");
-  }
-  const { units, total } = collected.value;
-  return {
-    entries: units.map(({ taken, start, end }) => ({
-      childIndex: taken.childIndex,
-      end,
-      node: taken.node,
-      start,
-    })),
-    paragraph,
-    total,
-  };
-}
-
-export function getAcceptedOffset(
-  point: AcceptedPoint,
-  map: AcceptedMap,
-): number | null {
-  let offset = 0;
-  const children = map.paragraph.getChildren();
-  for (let childIndex = 0; childIndex < children.length; childIndex += 1) {
-    if (childIndex > point.childIndex) {
-      break;
-    }
-    const child = children[childIndex];
-    if (child === undefined) {
-      return null;
-    }
-    if (childIndex === point.childIndex) {
-      if (point.node === null) {
-        return offset;
-      }
-      if (!$isTextNode(child) || child.getKey() !== point.node.getKey()) {
-        return null;
-      }
-      return offset + point.offset;
-    }
-    if ($isTextNode(child)) {
-      offset += child.getTextContentSize();
-    }
-  }
-  return point.node === null &&
-    point.childIndex === map.paragraph.getChildrenSize()
-    ? offset
-    : null;
 }
 
 function buildAcceptedSpan(
@@ -1099,16 +883,18 @@ export function getTargetSpanNodes(
   target: AcceptedRangeTarget | ProposalRangeTarget,
 ): TextNode[] {
   if (target.kind === "proposal-range") {
-    const mapped = proposalMapOf(target);
+    const mapped = proposalMapOf(target, inspectProposalGroup);
     if (mapped.status !== "ready") return [];
-    return mapped.value.entries
-      .filter((entry) => entry.start < target.end && entry.end > target.start)
-      .map((entry) => entry.node);
+    return overlappingSlices(
+      mapped.value.entries,
+      target.start,
+      target.end,
+    ).map((slice) => slice.node);
   }
   const map = buildAcceptedMap(target.paragraph);
-  return map.entries
-    .filter((entry) => entry.start < target.end && entry.end > target.start)
-    .map((entry) => entry.node);
+  return overlappingSlices(map.entries, target.start, target.end).map(
+    (slice) => slice.node,
+  );
 }
 
 /** Split span edges and return the span's live text nodes. */
@@ -1116,22 +902,16 @@ export function isolateTargetSpanNodes(
   target: AcceptedRangeTarget | ProposalRangeTarget,
 ): TextNode[] {
   const { start, end } = target;
-  let entries: readonly (ProposalMapEntry | AcceptedMapEntry)[];
   if (target.kind === "proposal-range") {
-    const mapped = proposalMapOf(target);
+    const mapped = proposalMapOf(target, inspectProposalGroup);
     if (mapped.status !== "ready") return [];
-    entries = mapped.value.entries;
-  } else {
-    entries = buildAcceptedMap(target.paragraph).entries;
+    return isolateSpanNodes(mapped.value.entries, start, end);
   }
-  return entries
-    .filter((entry) => entry.start < end && entry.end > start)
-    .map((entry) => {
-      const localStart = Math.max(start - entry.start, 0);
-      const localEnd = Math.min(end - entry.start, entry.end - entry.start);
-      const parts = entry.node.splitText(localStart, localEnd);
-      return parts[localStart === 0 ? 0 : 1]!;
-    });
+  return isolateSpanNodes(
+    buildAcceptedMap(target.paragraph).entries,
+    start,
+    end,
+  );
 }
 
 /** Fragment-local classification behind the seam: ownership plus offsets. */
@@ -1209,24 +989,6 @@ export function inspectFragmentSelection(): Preparation<FragmentSelection> | nul
       backward: selection.isBackward(),
     },
   };
-}
-
-/**
- * Fresh proposal map for one target at use time. Owners never carry maps;
- * every span operation rebuilds from the target's paragraph and wrappers.
- */
-export function proposalMapOf(target: {
-  paragraph: ParagraphNode;
-  wrappers: readonly ReviewElementNode[];
-}): Preparation<ProposalMap> {
-  const first = target.wrappers[0];
-  const last = target.wrappers.at(-1);
-  if (first === undefined || last === undefined)
-    return refusal(
-      "invalid-structural-target",
-      "The pending proposal has no live content.",
-    );
-  return buildProposalMap(target.paragraph, first, last);
 }
 
 export type StructuralPosition = Readonly<{
